# Improve BinaryHeap performance