@tavianator 2020-11-09
I don't know if this makes a difference performance-wise (or even readability-wise), but you could avoid the comparisons with `end - 1` by computing the right child by default instead of the left (`child = 2 * pos + 2`).