@SkiFire13 2020-11-09
@tavianator I didn't think of that, but I wouldn't expect it to make much of a difference (the compiler stores `end-1` in a register before the loop so no other calculation is required inside). I benched it anyway, and surprising it even performs worse.

|name                 |left child ns/iter|right child ns/iter|diff ns/iter  |diff %   |speedup |
|---------------------|------------------|-------------------|--------------|---------|--------|
|find_smallest_1000   | 260,098          | 268,056           |      +7,958  |  +3.06% | x 0.97 |
|from_vec             | 473,934          | 533,500           |     +59,566  | +12.57% | x 0.89 |
|into_sorted_vec      | 304,275          | 365,226           |     +60,951  | +20.03% | x 0.83 |
|pop                  | 373,778          | 392,372           |     +18,594  |  +4.97% | x 0.95 |

I also tested [a reduced version on godbolt](https://godbolt.org/z/qj1xfP), and looks like LLVM is doing a worse job at optimizing `sift_down_range`'s loop (13 vs 16 instructions for a `u64`, that's like 23% more instructions!) which might explain the slowdown, in particular `hole.get(child - 1)` is not optimized as good as `hole.get(child + 1)` previously was.