@tavianator 2020-11-09
Huh, thanks for checking.  That codegen is kinda odd.