@SkiFire13 2020-11-09
I've added the safety comment, I hope it's clear enough. I'm tempted to document other `unsafe` usages as well, but that should probably have its own PR.

r=KodrAus