@kennytm 2020-11-12
rust-lang/cargo#8517 cc @jyn514